requirements.txt
Flask==2.3.2
Werkzeug==2.3.6
PyPDF2==3.0.1
SQLAlchemy==2.0.15
pymysql==1.0.3
python-dotenv==1.0.0
openai==0.27.8